requirements.txt
# 核心依赖
python>=3.8

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
geopandas>=0.10.0
shapely>=1.8.0
pyproj>=3.2.0

# 空间分析
rasterio>=1.2.0
fiona>=1.8.0
rtree>=0.9.0

# 机器学习
scikit-learn>=1.0.0
scipy>=1.7.0

# 可视化
folium>=0.12.0
matplotlib>=3.5.0
plotly>=5.5.0
branca>=0.5.0

# Web框架
flask>=2.0.0  # 用于构建Web应用
dash>=2.0.0   # 用于交互式仪表板

# API调用
requests>=2.26.0

# 工具
jupyter>=1.0.0
ipython>=8.0.0
pyyaml>=6.0.0
tqdm>=4.62.0

# 开发工具
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0